+++ geode-core/src/main/java/org/apache/geode/management/internal/beans/MetricsCalculator.java
@@ -129,7 +129,7 @@ public class MetricsCalculator {
     float p = (float) Math.pow(10, Rpl);
     Rval = Rval * p;
     float tmp = Math.round(Rval);
-    return (float) tmp / p;
+    return tmp / p;
   }
